backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
openai>=1.10.0
json-repair>=0.28.0
httpx>=0.26.0
